[Libomptarget] Refactor Remote Plugin to Support UCX
The remote offloading plugin currently hardwires gRPC as the sole
backend communication library. In order to support UCX as well as gRPC,
these refactoring changes were necessary. In addition, this fixes a bug
where the default value of `__tgt_rtl_number_of_devices` returned -1
instead of 0 when the plugin could not connect to the server.