[Libomptarget][WIP] Introduce VGPU Plugin
This patch introduces a virtual GPU (x86) plugin. This allows for the
emulation of the GPU environment on the host. This re-uses the same
execution model, compilation paths, runtimes as a physical GPU. The
number of threads, warps, and CTAs are set through the environment
variables `VGPU_{NUM_THREADS,NUM_WARPS,WARPS_PER_CTA}` respectively.

Known Bugs (hence WIP):
- In the rebase from LLVM 12, larger applications started segfaulting.
  Small programs still work with this patch.
- The virtual GPU should be able to execute kernels asynchronously using
  the streams - but there is an unknown lifetime issue around the `ffi_call` that
  prevents the removal of the await after the `scheduleAsync` call.

Differential Revision: https://reviews.llvm.org/D113359